+++ libcontainer/stats_linux.go
@@ -1,10 +1,13 @@
 package libcontainer
 
-import "github.com/opencontainers/runc/libcontainer/cgroups"
-import "github.com/opencontainers/runc/libcontainer/intelrdt"
+import (
+	"github.com/opencontainers/runc/libcontainer/cgroups"
+	"github.com/opencontainers/runc/libcontainer/intelrdt"
+	"github.com/opencontainers/runc/types"
+)
 
 type Stats struct {
-	Interfaces    []*NetworkInterface
+	Interfaces    []*types.NetworkInterface
 	CgroupStats   *cgroups.Stats
 	IntelRdtStats *intelrdt.Stats
 }
